requirements.txt
boolean.py == 4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1     
pandas==2.2.2
PTable==0.9.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabulate==0.9.0
truth-table-generator==2.0.0
tzdata==2024.1
Flask==2.0.1
Werkzeug==2.2.2